* tree-vectorizer.h: Move declaration of vect_free_slp_tree under the 	relevant comment. 	* tree-vect-analyze.c (vect_analyze_data_ref_access): Fix comments 	and unify to if-stmts into one.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/autovect-branch@126074 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.dg/vect/slp-21.c
@@ -9,7 +9,7 @@
 int
 main1 ()
 {
-  int i;
+  unsigned short i;
   unsigned short out[N*8], out2[N*8], b0, b1, b2, b3, b4, a0, a1, a2, a3, b5;
   unsigned short in[N*8];
